warnings: fix up new warnings

+++ mm/debug.c
@@ -318,7 +318,7 @@ __setup("vm_debug", setup_vm_debug);
 void page_init_poison(struct page *page, size_t size)
 {
 	if (page_init_poisoning)
-		memset(page, PAGE_POISON_PATTERN, size);
+		memset(page, (u8)PAGE_POISON_PATTERN, size);
 }
 EXPORT_SYMBOL_GPL(page_init_poison);
 #endif		/* CONFIG_DEBUG_VM */
